% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{hello_rpkg}
\alias{hello_rpkg}
\title{Hello rpkg}
\usage{
hello_rpkg(rpkg)
}
\arguments{
\item{rpkg}{package name}
}
\value{
rpkg character
}
\description{
Hello rpkg
}
\examples{
hello_rpkg("rpkg.template")
}
