% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoir-package.R
\docType{package}
\name{pandoir-package}
\alias{pandoir}
\alias{pandoir-package}
\title{pandoir: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: William Song \email{william_swl@163.com}

}
\keyword{internal}
