% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{url2doi}
\alias{url2doi}
\title{get doi from url in parallel}
\usage{
url2doi(urls, retry = 5)
}
\arguments{
\item{urls}{urls}

\item{retry}{the time of retries}
}
\value{
dois
}
\description{
get doi from url in parallel
}
\examples{
future::plan(future::multisession)
url <- "https://www.nature.com/articles/s41586-022-05644-7"
url2doi(url)

}
