% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{url2doi_}
\alias{url2doi_}
\title{get doi from url}
\usage{
url2doi_(url)
}
\arguments{
\item{url}{url}
}
\value{
doi
}
\description{
get doi from url
}
\examples{
url <- "https://www.nature.com/articles/s41586-022-05644-7"
url2doi_(url)

}
